Provide more useful labels for our use-case
Adds a number of io.k8s.* and io.openshift.* tags that would make the
image metadata more useful for the OCP use-case. Adds a run label
that gives the idea how the container might be ran in practice (well,
probably in debugging, because in practice you'd only use the image
with an operator).