Use ubi8/ubi-minimal as base
Instead of using Fedora and installing the scanner from Fedora,
let's use ubi8/ubi-minimal as the base. Because the RHEL packages
are not adequate yet (missing oscap-chroot, missing some required
patches), we also need to provide a RHEL-8 COPR repo with a rebuild
of the container.

Finally, because there is neither yum nor dnf on the minimal image,
install the packages with microdnf.